Notes button:
The intention is instead of having a separate notes page,
I have added a notes button on the dashboard page.
This button will take the user to the notes page.
The notes page will have a button to go back to the dashboard page.

It's supposed to work like that but it's not 100% perfect, but at
least it's get the job done for now.